refactor(self_improve_executor): split tests.rs (#1266)
tests.rs 504 -> 264 LOC; new tests_extra.rs (~285 LOC).
